+++ test/units/vftr_setup_test_1.sh
@@ -1,11 +1,13 @@
 #!/bin/bash
 set -x
-ref_out_dir=ref_output
-ref_in_dir=ref_input
+ref_out_dir=${srcdir}/ref_output
+ref_in_dir=${srcdir}/ref_input
 testname=vftr_setup_test_1
 outfile=$testname.out
 
-rm -f $outfile
+if [ -f $outfile ] ; then
+   rm -f $outfile
+fi
 
 ./test_vftrace $testname
 diff $ref_out_dir/$outfile $outfile
